Refactoring: 파츠를 각각 분리 후 dto를 통한 module화
