--- mission5_dtos/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-# Keep environment variables out of version control
-.env
-
-/generated/prisma
