--- mission5_dtos/src/app.ts
@@ -1,31 +1,0 @@
-import express from "express";
-import cookieParser from "cookie-parser";
-import router from "./routers/index";
-import { PORT } from "./lib/constants";
-import passport from "./lib/passport/index";
-import cors from "cors";
-import { errorHandler } from "./middlewares/errorHandler";
-
-const app = express();
-
-app.use(express.urlencoded({ extended: true }));
-
-app.use(express.json());
-app.use(cookieParser());
-const corsOptions = {
-  //CORS 설정 및 whitelist로 'http://localhost:3000'을 설정
-  origin: "http://localhost:3000",
-  optionSuccessStatus: 200,
-};
-
-app.use(cors(corsOptions));
-app.use(passport.initialize());
-app.use(router);
-app.use("/download", express.static("uploads"));
-
-
-app.use(errorHandler);
-
-app.listen(PORT, () => {
-  console.log(`Server is running on http://localhost:${PORT}`);
-});
