--- mission5_dtos/src/routers/articleRouter.ts
@@ -1,14 +1,0 @@
-import express from "express";
-import passport from "../lib/passport";
-import { ArticleController } from "../controllers/articleController";
-
-const router = express.Router();
-const controller = new ArticleController();
-
-router.post("/", passport.authenticate("access-token", { session: false }), controller.create);
-router.get("/", passport.authenticate("access-token", { session: false }), controller.list);
-router.get("/:id", passport.authenticate("access-token", { session: false }), controller.detail);
-router.patch("/:id", passport.authenticate("access-token", { session: false }), controller.update);
-router.delete("/:id", passport.authenticate("access-token", { session: false }), controller.delete);
-
-export default router;
